+++ step/build_id_getter_test.go
@@ -1,4 +1,4 @@
-package main
+package step
 
 import (
 	"sort"
